requirements.txt
asgiref==3.6.0
Django==4.1.4
django-taggit==3.1.0
Pillow==9.3.0
sqlparse==0.4.3
